+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Versioning
 
 This project does not follow a versioning standard. Versions are crafted after the dates; for example, the version 17.7.25 was released on July, 25th in 2017
+
+## X.X.X
+### Fixed
+-   Made sure the reserved value matches the gross value of the order to avoid problems finalizing.
+
 ## 19.10.17
 ### Added
 - Hash validation to push requests.